Implement low latency heartbeats over TCP connections by only verifying if any data has been sent or received between a heartbeat request and the timeout. If any data has been sent (possibly the heartbeat request), it implies that a TCP ACK must have been received from the remote party which implies that the remote party is alive. If any data has been received, it implies that the remote party is alive.

+++ src/EventStore.Core/Messages/TcpMessage.cs
@@ -32,9 +32,9 @@ namespace EventStore.Core.Messages {
 				get { return TypeId; }
 			}
 
-			public readonly int MessageNumber;
+			public readonly long MessageNumber;
 
-			public Heartbeat(int messageNumber) {
+			public Heartbeat(long messageNumber) {
 				MessageNumber = messageNumber;
 			}
 		}
@@ -46,9 +46,9 @@ namespace EventStore.Core.Messages {
 				get { return TypeId; }
 			}
 
-			public readonly int MessageNumber;
+			public readonly long MessageNumber;
 
-			public HeartbeatTimeout(int messageNumber) {
+			public HeartbeatTimeout(long messageNumber) {
 				MessageNumber = messageNumber;
 			}
 		}
